fix(identity): resolve display name from server with memoized cache and invalidation
